Remove redundant syntax rule for Product
The syntax definition for a product of sorts consist of two syntax rules, but I don't see why it is just one rule. In the current construction additional parentheses are needed to use Product. For example, now one has to write Product((Bool(), [Bool()])), but with this change one would have to write Product(Bool(), [Bool()]).