Syntax rule for conditional data expression
The conditional data expression `if(b, c, d)` meaning "if b holds then c else d" is missing in the syntax definition. See the definition of this mapping [here](https://mcrl2.org/web/user_manual/language_reference/data.html#predefined-mappings).

As this is a circumfix operator I expect that it is not necessary to define priorities for this operator.